requirements.txt
flask==3.0.3
pdfminer.six==20240706
nltk==3.9.1
torch>=2.2.0
openai>=1.30.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
filelock==3.19.1
fsspec==2025.7.0
huggingface-hub==0.34.4
idna==3.10
packaging==25.0
PyYAML==6.0.2
requests==2.32.5
tokenizers==0.21.4
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0